// src/domain/models/ClientesModel.ts
export interface ClientesModel {
    IdCliente:       number;
    TipoDoc:         string;
    NroDoc:          string;
    NombreCompleto:  string;
    FechaNac:        Date;
    Direccion:       string;
    email:           string;
    password:        string;
    Telefono:        string;
    Estado:          boolean;    
    IdUbigeo:        number;
}